% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_new_api.R
\name{api_forecast}
\alias{api_forecast}
\title{change town forcast api data of format.}
\usage{
api_forecast(cat = "T3H", url = url[1:2])
}
\description{
The function is to cange downloaded files format
}
\details{
@param category= POP,PTY, REH,SKY,T3H,UUU,VEC,VVV,WSD,WAV
 @param url = api url
 @return
 @example
 @export
}
