% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack2.R
\name{pack2}
\alias{pack2}
\title{install.packages & library}
\usage{
pack2(package)
}
\description{
mom_gamma function will calculate mom estimate for the shape parameter "k" and the scale parameter "theta" by using given sample.
}
\details{
@param package is package names.
 @return package & library
 @examples #' pack2(package=c('sp','plyr'))
 @export
}
