% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_new.R
\name{town_new}
\alias{town_new}
\title{change town forcast data of format.}
\usage{
town_new(dir1)
}
\description{
The function is to cange downloaded files format
}
\details{
@param dir1 = directory of save files
 @param City_index is an index to the list of citydata2.For example, if the city_index=1 is download by Seoul data.

@return
 @examples dir1='D:/dir/data'
 town_new(dir1=dir1)
 @export
}
